client_api/requirements.txt
distro==1.9.0
fastapi==0.115.13
paramiko==3.5.1
psutil==7.0.0
PyJWT==2.10.1
python-multipart==0.0.20
requests==2.32.3
uvicorn==0.34.3
websockets==15.0.1
